Backend code refactoring and token cache implementation
