+++ backend/formatting.py
@@ -1,125 +1,8 @@
-from pydantic_settings import BaseSettings
-from fastapi import HTTPException
-from pymongo.mongo_client import MongoClient
-from pymongo.server_api import ServerApi
-from bson.binary import UuidRepresentation
 import json
 import re
-import traceback
 from io import StringIO
 from html.parser import HTMLParser
-import uuid
-import datetime
-
-
-def getMongoConnectionUri(settings: BaseSettings):
-    mongo_conn_start = "mongodb+srv://"
-    mongo_conn_end = "@learntocodecluster0.pvsvjov.mongodb.net/?retryWrites=true&w=majority"
-    return mongo_conn_start + settings.mongo_db_user + ":" + settings.mongo_db_pass + mongo_conn_end
-
-
-def createAndStoreUserToken(settings: BaseSettings):
-    uri = getMongoConnectionUri(settings)
-    client = MongoClient(uri, server_api=ServerApi('1'), uuidRepresentation='standard')
-
-    try:
-        db = client.learntocodedb
-        sessions = db.sessions
-
-        newToken = uuid.uuid4()
-        currentTime = datetime.datetime.now(tz=datetime.timezone.utc)
-        tokenLife = datetime.timedelta(weeks=1)
-        newSession = {
-            "token": newToken,
-            "expirationDate": currentTime + tokenLife
-        }
-        sessions.insert_one(newSession)
-
-        return newToken
-    except Exception as e:
-        print(e)
-        raise HTTPException(status_code=500, detail="There was an issue authenticating your session.")
-
-
-def validateAuthHeader(auth_header: str, settings: BaseSettings):
-
-    if(settings.api_auth_enabled):
-        uri = getMongoConnectionUri(settings)
-        client = MongoClient(uri, server_api=ServerApi('1'), uuidRepresentation='standard')
-        # Need to check MongoDB instance for token existence and validity (not expired)
-        try:
-            db = client.learntocodedb
-            sessions = db.sessions
-
-            queryParameters = {
-                "token": uuid.UUID(auth_header),
-                "expirationDate": { "$gt": datetime.datetime.now(tz=datetime.timezone.utc) }
-            }
-
-            session = sessions.find_one(queryParameters)
-
-            print(session)
-
-            if session == None:
-                raise HTTPException(status_code=401, detail="You are not authorized to call this endpoint. Please log in again.")
-        except Exception as e:
-            print(e)
-            raise HTTPException(status_code=500, detail="There was an issue authenticating your session.")
-    
-    return True
-
-
-def validateCode(code: str, auth_header: str, settings: BaseSettings):
-    validateAuthHeader(auth_header, settings)
-
-    try:
-        codeObject = compile(code, "User supplied code", "exec")
-    except Exception as e:
-        eType = type(e).__name__
-        userErrorString = ""
-        if eType == 'SyntaxError':
-            traceStrList = traceback.format_exception(e)
-            strLen = len(traceStrList)
-            if strLen >= 3:
-                userErrorString = traceStrList[strLen-3] + traceStrList[strLen-2] + traceStrList[strLen-1]
-
-        errorObject = {
-            "errorType": eType,
-            "errorMessage": userErrorString
-        }
-
-        print("I have caught an error during compilation!")
-
-        raise HTTPException(status_code=400, detail=errorObject)
-
-    return None
-
-
-def validateAndParsePrompts(requestedSection: str, id: int, auth_header: str,  prompts: any, settings: BaseSettings):
-    validateAuthHeader(auth_header, settings)
-            
-    section = prompts.get(requestedSection, None)
-    if(section == None): 
-        raise HTTPException(status_code=400, detail="Invalid section provided.")
-    
-    requestedPrompts = []
-    
-    if(id != -1):
-        exercisePrompt = section.get(id, None)
-        if(exercisePrompt == None):
-            raise HTTPException(status_code=400, detail="Exercise ID provided is invalid for given section")
-        
-        requestedPrompts.append((id, exercisePrompt))
-    else:
-        # Empty dictionaries evaluate to false
-        if(not section):
-            raise HTTPException(status_code=400, detail="Section provided is empty")
-        
-        for key in section:
-            requestedPrompts.append((key, section.get(key)))
-        
-    return requestedPrompts
-
+from fastapi import HTTPException
 
 class MLStripper(HTMLParser):
     def __init__(self):
