+++ backend/test/variables/sample_formatted_completion_1.json
