+++ backend/test/variables/sample_formatted_completion_2.json
